#195: Stabilise Compose API usages
* Replace access to Compose internal `LocalWindow` with custom detection of the local window
* Completely remove access to `loadStateAndComposeForHotReload` and replace the all uses with 'reset composition'
