+++ hot-reload-devtools/src/test/kotlin/org/jetbrains/compose/devtools/tests/ui/ExpandedSidecarUiTest.kt
@@ -134,9 +134,9 @@ class ExpandedSidecarUiTest : SidecarBodyUiTest() {
     @Test
     fun `test - actions`() = runSidecarUiTest {
         onAllNodesWithTag(Tag.ActionButton)
-            .assertCountEquals(4)
+            .assertCountEquals(3)
             .filter(hasClickAction())
-            .assertCountEquals(4)
+            .assertCountEquals(3)
     }
 
     @Test
