+++ hot-reload-runtime-jvm/src/main/kotlin/org/jetbrains/compose/reload/jvm/resetComposition.kt
@@ -7,7 +7,6 @@ package org.jetbrains.compose.reload.jvm
 
 import kotlinx.coroutines.flow.update
 
-internal fun cleanComposition() {
+internal fun resetComposition() {
     hotReloadState.update { state -> state.copy(key = state.key + 1) }
-    retryFailedCompositions()
 }
